+++ grpcutil/util.go
@@ -1,4 +1,4 @@
-package grpc
+package grpcutil
 
 import (
 	"github.com/gogo/status"
